// backend/src/modules/notifications/types/notification-type.enum.ts
export enum NotificationType {
  BOOKING_CREATED = 'BOOKING_CREATED',
  BOOKING_CONFIRMED = 'BOOKING_CONFIRMED',
  BOOKING_CANCELLED = 'BOOKING_CANCELLED',
  PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
  PAYMENT_FAILED = 'PAYMENT_FAILED',
  CHECK_IN_COMPLETED = 'CHECK_IN_COMPLETED',
  CHECK_OUT_COMPLETED = 'CHECK_OUT_COMPLETED',
  REVIEW_CREATED = 'REVIEW_CREATED',
  SYSTEM = 'SYSTEM',
}